+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For details about compatibility between different releases, see the **Commitment
 ### Added
 
 - More password validation rules in the user management form in the Console.
+- Support for sending end device uplinks using the CLI (see `ttn-lw-cli end-devices simulate-uplink` command).
 
 ### Changed